Moved playerLayer to NewTabPageBackgroundView
